# docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Clientes API - Projeto Final Java WebDeveloper",
        "description": "API REST para gerenciamento de clientes e seus endere\u00e7os.",
        "contact": {
            "name": "COTI Inform\u00e1tica"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "cliente-controller"
                ],
                "operationId": "getAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "cliente-controller"
                ],
                "operationId": "put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditarClienteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "cliente-controller"
                ],
                "operationId": "post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CriarClienteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "cliente-controller"
                ],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cliente-controller"
                ],
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EditarClienteRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 8
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "cpf": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^\\d{11}$"
                    },
                    "dataNascimento": {
                        "type": "string",
                        "format": "date"
                    },
                    "endereco": {
                        "$ref": "#/components/schemas/EditarEnderecoRequest"
                    }
                },
                "required": [
                    "cpf",
                    "dataNascimento",
                    "email",
                    "endereco",
                    "id",
                    "nome"
                ]
            },
            "EditarEnderecoRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "logradouro": {
                        "type": "string",
                        "minLength": 1
                    },
                    "complemento": {
                        "type": "string"
                    },
                    "numero": {
                        "type": "string",
                        "minLength": 1
                    },
                    "bairro": {
                        "type": "string",
                        "minLength": 1
                    },
                    "cidade": {
                        "type": "string",
                        "minLength": 1
                    },
                    "uf": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[A-Za-z]{2}$"
                    },
                    "cep": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^\\d{8}$"
                    }
                },
                "required": [
                    "bairro",
                    "cep",
                    "cidade",
                    "logradouro",
                    "numero",
                    "uf"
                ]
            },
            "CriarClienteRequest": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 8
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "cpf": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^\\d{11}$"
                    },
                    "dataNascimento": {
                        "type": "string",
                        "format": "date"
                    },
                    "endereco": {
                        "$ref": "#/components/schemas/EnderecoRequest"
                    }
                },
                "required": [
                    "cpf",
                    "dataNascimento",
                    "email",
                    "endereco",
                    "nome"
                ]
            },
            "EnderecoRequest": {
                "type": "object",
                "properties": {
                    "logradouro": {
                        "type": "string",
                        "minLength": 1
                    },
                    "complemento": {
                        "type": "string"
                    },
                    "numero": {
                        "type": "string",
                        "minLength": 1
                    },
                    "bairro": {
                        "type": "string",
                        "minLength": 1
                    },
                    "cidade": {
                        "type": "string",
                        "minLength": 1
                    },
                    "uf": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^[A-Za-z]{2}$"
                    },
                    "cep": {
                        "type": "string",
                        "minLength": 1,
                        "pattern": "^\\d{8}$"
                    }
                },
                "required": [
                    "bairro",
                    "cep",
                    "cidade",
                    "logradouro",
                    "numero",
                    "uf"
                ]
            }
        }
    }
}